Got voting interaction working

+++ app/models/user.rb
@@ -26,11 +26,19 @@ class User < ActiveRecord::Base
   def upvote_story(story)
     
     user_vote = self.votes.find_by_story_id(story.id)
-    
     if user_vote
-      vote = user_vote.destroy
-      point_change = -1
+      logger.info "*** VOTE"
+      if user_vote.up_or_down == "down"
+        logger.info "***     IS DOWN"
+        vote = user_vote.update_attributes(:up_or_down => :up)
+        point_change = 2
+      else 
+        logger.info "***     IS UP"
+        vote = user_vote.destroy
+        point_change = -1
+      end
     else
+      logger.info "*** NO VOTE"
       vote = self.votes.create(:user_id => self.id, :story_id => story.id, :up_or_down => :up)
       point_change = 1
     end
@@ -46,11 +54,19 @@ class User < ActiveRecord::Base
   def downvote_story(story)
     
     user_vote = self.votes.find_by_story_id(story.id)
-    
     if user_vote
-      vote = user_vote.destroy
-      point_change = 1
+      logger.info "*** VOTE"
+      if user_vote.up_or_down == "up"
+        logger.info "***     IS UP"
+        vote = user_vote.update_attributes(:up_or_down => :down)
+        point_change = -2
+      else 
+        logger.info "***     IS DOWN"
+        vote = user_vote.destroy
+        point_change = 1
+      end
     else
+      logger.info "*** NO VOTE"
       vote = self.votes.create(:user_id => self.id, :story_id => story.id, :up_or_down => :down)
       point_change = -1
     end
